Changed int to float Problem, implemented strategies for float and polynomials problem. Implemented Generations and TimeElapsed StopStrategy.

+++ src/main/java/model/factory/PolynomialProblemFactory.java
@@ -11,6 +11,6 @@ public class PolynomialProblemFactory implements IProblemFactory {
 
 
     public IProblem createProblem() {
-        return new PolynomialProblem(new int[]{1,2,3});
+        return new PolynomialProblem(new float[]{1,2,3});
     }
 }
